Big Int instead of int

+++ Assignment4/config.php
@@ -37,7 +37,7 @@ if(!$conn){
    $sql = "CREATE TABLE FuelQuote (
     id INT(6) UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     username VARCHAR(30) NOT NULL,
-    gallonsRequested INT(255) NOT NULL,
+    gallonsRequested BIGINT NOT NULL,
     deliveryAddress VARCHAR(100) NOT NULL,
     deliveryDate DATE NOT NULL,
     dollarsPerGallon DOUBLE(255, 2) UNSIGNED  NOT NULL,
